Chunking-project/requirements.txt
chromadb==1.5.5
faiss-cpu==1.13.2
groq==1.1.2
numpy==2.4.3
pypdf==6.9.2
python-docx==1.2.0
sentence-transformers==5.3.0
streamlit==1.55.0
